requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
dashscope==1.13.6
loguru==0.7.2
sqlalchemy==2.0.23
pg8000==1.30.2  # 纯Python实现的PostgreSQL驱动，替代psycopg2-binary
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6
